Rename validator and fix tests
